requirements.txt
tornado==4.4.1
pika==0.10.0
nose==1.3.7
wheel==0.24.0
coverage==4.2
sqlalchemy-migrate==0.10.0
PyMySQL==0.7.6
configobj==5.0.6
python-dateutil
requests == 2.18.1
mock == 2.0.0